Added bloc architecture, image loaded and displayed to the user
The file picker filters images only with .jpg .png . After the user has picked a file the file path gets displayed in a text field with the size of the image being automatically inputted for the user. After this the image is displayed on the left. Now working on connecting the front and back